+++ src/store/todoContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+export const todoContext = createContext(null)
+
+export default todoContext
